requirements.txt
bleak
automower_ble
paho.mqtt